Display 'Unsupported event' instead of writing to the console
While there is a list of possible event types, I only implemented
those that are in my activity stream at this point, as these have
data that I can use to test.

Thus, I want to immediately see any unsupported event types so I
can support it, instead of writing to a console.

Signed-off-by: Ealden Esto E. Escanan <ealden@gmail.com>