Display unsupported event message on events that does not match our conditions
While we try to support different events, each event may have subevents
that we do not cover.  For example, create event has subevents for
creating branches and creating repositories, and perhaps more that we
don't handle.  For these scenarios we display unsupported event and log
the necessary info to the console.

Signed-off-by: Ealden Esto E. Escanan <ealden@gmail.com>